Add extra configuration for map layer
For the map layer of type wms is added extra configuration wmsFormat,
wmsVersion.
Also the wmsUrl is renamed to url because the url can be used also from
other layer types.
There are added also two extra parameters securityUser and
securityPassword. At the moment, these are not used but it might be needed
in the future.
The GetMapDefinition is moved from security free web service spatial to
search.